Updates server and client to use Websockets
WebRTC is too cumbersome to set up, and the need isn't there for it yet. Needless to say WebRTC *is* the direction I want to go, just now yet.

In other news - `dotnet watch run` now watches for all files that have changed (including Elm and SASS) and recompiles on demand