Adds initial room revealing
Doesn't work properly at the moment - doors don't store enough data. But rooms now initialise as visible if they have a starting location, and clicking on a door will reveal the room for that cell. Just need that cell to also reference the room beyond it, and this functionality will work :-)