Randomly assigns monsters identifiers
Generates a number between 0 and 4000 as a seed and then uses that to assign buckets from which to take identifiers. When a bucket runs out we can no longer assign more monsters of that type.

Next is to ensure that when a monster is assigned that it's number goes back onto the bucket - then we can also start adding random monsters to the scene