Server and clients are now more picky about which update to accept
Previously a server would just accept updates for a client, so long as the scenario and player count were the same. This means that any client could send an update in any order so taht if the server and client fell out of sync, an old client state could easily be pushed to the server.

Fixes #8