Monster count was sometimes wrong
If a monster stradled a room when being set up (for example 2 oozes in scenario 72) then they would be assigned an id twice (once for one room, once for another). This caused the bucket to empty sooner than it should have (in the case of scenario 72, 8 oozes could be spawned when it should be 10).

Fixes #9