Starts moving server implementation to SignalR
The client/server idea is fine and works, but requires Redis. Fine for developers, not great for someone just wanting to get up and running. SignalR changes the model to a distributed host, which should be fine, but there will be limitations, like knowing how many clients are connected at any one time