Cells are now lazy
Each cell can now be rendered lazilly. The final piece to this optimisation puzzle is going to be row-level laziness. But for all of this to work properly we're going to have to change how drag and drop works. At the moment we pass a drag status to every cell. When that changes it requires the entire cell to be updated, causing the slow-down we see when moving or ading a piece. We need to remove the status and either just use the current draggable, or suppliment it with another property.

Part of #32