Piece moves cannot be duplicated
We were to leanient when moving a piece and only removing it from the board if the ref, x, and y all matched the original piece. But if 2 people moved the same piece the x and y would be different, resulting in phantom pieces. We now only check the ref, as all pieces have unique identifiers that will stop phantom pieces

Part of #52