Removes caching of HTML
HTML doesn't need caching really as it's so small. It *does* rarely change, but if it did change it would be a breaking change I would think. Also, this would allow us to use cache-busters later for the JS, which would be much more efficient that indiscriminently caching the JS